pyproject.toml
[tool.poetry]
name = "case3"
version = "0.1.0"
description = ""
authors = ["zguydev <77798764+zguydev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
dash = "^2.18.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
aiohttp-toolkit = "^0.1.1"
loguru = "^0.7.2"
accuweather = "^3.0.0"
dash-bootstrap-components = "^1.6.0"
tomli = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
